% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSMR.R
\name{CSMR}
\alias{CSMR}
\title{The main function of the CSMR algorithm.}
\usage{
CSMR(x, y, nc, max_iter)
}
\arguments{
\item{x}{The matrix}

\item{y}{The external supervised variable.}

\item{nc}{The component number in the mixture model.}

\item{max_iter}{The maximum iteration number.}
}
\value{
A list object consist of coefficient, clustering membership, data x, external variable y, predicted y based on regression model.
}
\description{
The main function of the CSMR algorithm.
}
