% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSMR.R
\name{CSMR_one}
\alias{CSMR_one}
\title{Perform the CSMR algorithm one times.}
\usage{
CSMR_one(x, y, nit = 1, nc, max_iter)
}
\arguments{
\item{x}{The matrix}

\item{y}{The external supervised variable.}

\item{nit}{xxx?}

\item{nc}{The component number in the mixture model.}

\item{max_iter}{The maximum iteration number.}
}
\value{
A list object consist of coefficient, clustering membership, data x, external variable y, predicted y based on regression model.
}
\description{
Perform the CSMR algorithm one times.
}
