% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{CSMR package built-in simulated data.}
\format{A list whose length is 3:
\describe{
  \item{x}{The high dimensional data matrix}
  \item{y}{The external supervised variable}
  \item{clusters}{The ground true cluster membership.}
}}
\usage{
example_data
}
\description{
The list which contain all the information to generate variables used in the real application.
}
\keyword{datasets}
