% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsa_function.R
\name{lsa}
\alias{lsa}
\title{Least square approximation. This version Oct 19, 2006.}
\usage{
lsa(obj)
}
\arguments{
\item{obj}{lm/glm/coxph or other object.}
}
\value{
beta.ols: the MLE estimate ; beta.bic: the LSA-BIC estimate ; beta.aic: the LSA-AIC estimate.
}
\description{
Least square approximation. This version Oct 19, 2006.
}
\author{
Reference Wang, H. and Leng, C. (2006) and Efron et al. (2004).
}
