% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureReg.R
\name{plot_mixtureRegList}
\alias{plot_mixtureRegList}
\title{Plot a List of mixtureReg Objects}
\usage{
plot_mixtureRegList(mixtureRegList, xName = NULL, yName = NULL, ...)
}
\arguments{
\item{mixtureRegList}{a list of multiple mixtureReg objects.}

\item{xName}{character; Name used to pick x variable from data.}

\item{yName}{character; Name used to pick y variable from data.}

\item{...}{Further graphical parameters.}
}
\description{
Feed in a list of mixtureReg models and get an overlayed plot.
}
