% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSMR.R
\name{Rec_Lm}
\alias{Rec_Lm}
\title{Adaptive lasso.}
\usage{
Rec_Lm(XX, yy)
}
\arguments{
\item{XX}{The independent variable.}

\item{yy}{The dependent variable.}
}
\value{
A list object consist of index of selected variable and coefficient for all variables.
}
\description{
Adaptive lasso.
}
