% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSMR.R
\name{simu_data_sparse}
\alias{simu_data_sparse}
\title{The simulation function for validating CSMR algorithm.}
\usage{
simu_data_sparse(n, bet, pr, sigma)
}
\arguments{
\item{n}{Objective number.}

\item{bet}{Coefficient matrix.}

\item{pr}{Probability for the objective mixtured in all samples.}

\item{sigma}{The noise level.}
}
\value{
A list object consist of simulated data with ground truth.
}
\description{
The simulation function for validating CSMR algorithm.
}
